feat: expand View configuration options and add Aabb and Box support

+++ c/filament/c/Renderer.h
@@ -15,7 +15,7 @@ typedef struct FilaRendererFrameRateOptions {
     float headRoomRatio;
     float scaleRate;
     uint8_t history;
-    uint8_t interval;
+    float interval;
 } FilaRendererFrameRateOptions;
 
 typedef struct FilaRendererClearOptions {
